refactor: user리스트 관리 책임 분리
