+++ assignment/week1/src/main/java/org/example/Main.java
@@ -2,6 +2,7 @@ package org.example;
 
 import org.example.controller.BankSystem;
 import org.example.domain.BankMenu;
+import org.example.repository.CustomerRepository;
 
 public class Main {
     public static void main(String[] args) {
