+++ assignment/week1/src/main/java/org/example/domain/Account.java
@@ -1,4 +1,4 @@
-package org.example.repository;
+package org.example.domain;
 
 public class Account {
     private String accountNumber;
