+++ assignment/week1/src/main/java/org/example/domain/User.java
@@ -1,4 +1,6 @@
-package org.example.repository;
+package org.example.domain;
+
+import org.example.domain.Account;
 
 import java.util.List;
